+++ تقدیر/__init__.py
@@ -2,7 +2,7 @@ import os
 
 from எண்ணிக்கை import உரைக்கு as உ
 from .ذریعہ import ذریعہ
-from .کام import متاغیرات, اختیارہ_پانا, اختیارہ_رکھنا
+from .کام import متغیرات, اختیارہ_پانا, اختیارہ_رکھنا
 
 with open(os.path.join(os.path.split(__file__)[0], 'تبدیل.txt'), 'r', encoding='utf8') as تبدیل_کی_دستاویز:
     __تبدیل__ = உ(تبدیل_کی_دستاویز.read().strip(), 'latin', 'ار')
