--- تقدیر/ذرائع/جسان.py
@@ -1,48 +1,0 @@
-import json
-
-import pandas as pd
-from chardet import UniversalDetector
-from تقدیر.ذریعہ_نکتہ import ذریعہ_نکتہ
-from تقدیر import متاغیرات
-
-
-class جسان(ذریعہ_نکتہ):
-    """
-    جسان (.json) مسل سے کوائف پڑھتا ہیے۔
-    """
-
-    def __init__(خود, مسل, چوڑائی, طول, بلندی=None, خاکے=None, تبدل_ستون=None):
-        super().__init__(چوڑائی, طول, بلندی, خاکے, تبدل_ستون)
-
-        if isinstance(مسل, str):
-            ضابطہ = _ضابطہ_بندی(مسل)
-            with open(مسل, 'r', encoding=ضابطہ) as م:
-                خود.جسان = json.load(م)
-        else:
-            خود.جسان = مسل
-
-    def _کوائف_بنانا(خود, سے, تک, چوڑائی, طول, بلندی, خاکے):
-
-        ستون_تاریخ = خود._نام_ستون('تاریخ')
-
-        اعداد_جسان = pd.DataFrame(
-            data={
-                س: خود.جسان[خود._نام_ستون(س)] for س in متاغیرات if خود._نام_ستون(س) in خود.جسان
-            }, index=خود._اشاریہ_پاندس_بنانا(خود.جسان[ستون_تاریخ])
-        )
-        return اعداد_جسان
-
-
-def _ضابطہ_بندی(مسل):
-    آلہ = UniversalDetector()
-    with open(مسل, 'rb') as م:
-        for لکیر in م.readlines():
-
-            آلہ.feed(لکیر)
-
-            if آلہ.done:
-                break
-
-    آلہ.close()
-
-    return آلہ.result['encoding']
