+++ تقدیر/ذرائع/جمع.py
@@ -2,5 +2,5 @@ from تقدیر.ذریعہ import ذریعہ
 
 
 class جمع(ذریعہ):
-    def کوائف_پانا(خود, سے, تک, چوڑائی, طول, بلندی, خاکے='۸۔۵'):
+    def _کوائف_بنانا(خود, سے, تک, چوڑائی, طول, بلندی, خاکے):
         pass
