+++ تقدیر/ذرائع/سی_اس_وی.py
@@ -3,7 +3,7 @@ import csv
 import pandas as pd
 from chardet import UniversalDetector
 
-from تقدیر import متاغیرات
+from تقدیر import متغیرات
 from تقدیر.ذریعہ_نکتہ import ذریعہ_نکتہ
 from எண்ணிக்கை import எண்ணுக்கு as எ
 
@@ -17,8 +17,8 @@ class سی_اس_وی(ذریعہ_نکتہ):
 
     def _کوائف_بنانا(خود, سے, تک, چوڑائی, طول, بلندی, خاکے):
 
-        ستون_تاریخ = خود._نام_ستون('تاریخ')
-        ف_ستون = [(خود._نام_ستون(س), س) for س in متاغیرات]
+        ستون_تاریخ = خود._نام_عمودی_ستون('تاریخ')
+        ف_ستون = [(خود._نام_عمودی_ستون(س), س) for س in متغیرات]
 
         اعداد = {}
         تاریخیں = []
@@ -38,7 +38,7 @@ class سی_اس_وی(ذریعہ_نکتہ):
                         ق = எ(ل[س].strip())
                         اعداد[س_پا].append(ق)
 
-        return pd.DataFrame(اعداد, index=خود._اشاریہ_پاندس_بنانا(تاریخیں))
+        return pd.DataFrame(اعداد, index=خود._اشاریہ_پانڈا_بنانا(تاریخیں))
 
 
 def _ضابطہ_بندی(مسل):
