+++ تقدیر/ذریعہ_نکتہ.py
@@ -18,7 +18,7 @@ class ذریعہ_نکتہ(ذریعہ):
 
         return super().کوائف_پانا(سے, تک, چوڑائی, طول, بلندی, خاکے='۸۔۵')
 
-    def _نام_ستون(خود, ستون):
+    def _نام_عمودی_ستون(خود, ستون):
         try:
             return خود.تبدل_ستون[ستون]
         except KeyError:
