+++ تقدیر/کام.py
@@ -84,4 +84,4 @@ def تاریخ_بنانا(تاریخ):
     return pd.to_datetime(تاریخ).date()
 
 
-متاغیرات = ['بارش', 'شمسی_تابکاری', 'درجہ_حرارت_زیادہ', 'درجہ_حرارت_کم', 'درجہ_حرارت_اوسط']
+متغیرات = ['بارش', 'شمسی_تابکاری', 'درجہ_حرارت_زیادہ', 'درجہ_حرارت_کم', 'درجہ_حرارت_اوسط']
