+++ تقدیر/کوائف.py
@@ -3,7 +3,7 @@ from datetime import date
 
 import pandas as pd
 
-from تقدیر.کام import متاغیرات
+from تقدیر.کام import متغیرات
 
 
 class کوائف(object):
@@ -31,7 +31,7 @@ class کوائف(object):
         سے = سے or اعداد.index.min()
         تک = تک or اعداد.index.max()
 
-        خود.اعداد = pd.DataFrame(columns=متاغیرات, index=pd.period_range(سے, تک))
+        خود.اعداد = pd.DataFrame(columns=متغیرات, index=pd.period_range(سے, تک))
         if اعداد is not None:
             خود.اعداد = خود.اعداد.combine_first(اعداد)
         خود.اعداد = خود.اعداد[سے:تک]
